I'm dumb I hope I fixed it

+++ Code/save.js
@@ -78,5 +78,7 @@
       }
     }
 function clear(){
-     localStorage.clear();
+     if(confirm('You are about to reset your game. All unlocked locations and eggs will be removed. Proceed?')){
+          localStorage.clear();
+     }
 }
